chore: move types to its own package
