+++ e2e/_suites/fleet/stand-alone.go
@@ -11,11 +11,11 @@ import (
 
 	"github.com/cenkalti/backoff/v4"
 	elasticversion "github.com/elastic/e2e-testing/internal"
-	types "github.com/elastic/e2e-testing/internal"
 	"github.com/elastic/e2e-testing/internal/common"
 	"github.com/elastic/e2e-testing/internal/deploy"
 	"github.com/elastic/e2e-testing/internal/installer"
 	"github.com/elastic/e2e-testing/internal/kibana"
+	"github.com/elastic/e2e-testing/internal/types"
 	"github.com/elastic/e2e-testing/internal/utils"
 
 	"github.com/elastic/e2e-testing/internal/elasticsearch"
