+++ e2e/_suites/kubernetes-autodiscover/autodiscover_test.go
@@ -23,8 +23,8 @@ import (
 	"github.com/cucumber/godog/colors"
 	messages "github.com/cucumber/messages-go/v10"
 	apme2e "github.com/elastic/e2e-testing/internal"
-	types "github.com/elastic/e2e-testing/internal"
 	beats "github.com/elastic/e2e-testing/internal/beats"
+	"github.com/elastic/e2e-testing/internal/types"
 	log "github.com/sirupsen/logrus"
 	flag "github.com/spf13/pflag"
 	"go.elastic.co/apm"
