+++ internal/installer/elasticagent_rpm.go
@@ -9,11 +9,11 @@ import (
 	"fmt"
 	"strings"
 
-	types "github.com/elastic/e2e-testing/internal"
 	"github.com/elastic/e2e-testing/internal/beats"
 	"github.com/elastic/e2e-testing/internal/common"
 	"github.com/elastic/e2e-testing/internal/deploy"
 	"github.com/elastic/e2e-testing/internal/kibana"
+	"github.com/elastic/e2e-testing/internal/types"
 	log "github.com/sirupsen/logrus"
 	"go.elastic.co/apm"
 )
